fix(layout): update to include correct imports
