+++ packages/layout/src/spacing.js
@@ -5,6 +5,8 @@
  * LICENSE file in the root directory of this source tree.
  */
 
+import { miniUnits } from './mini-unit';
+
 // Spacing
 export const spacing01 = miniUnits(0.25);
 export const spacing02 = miniUnits(0.5);
